Add timeFormat and fieldOptions to default new cell
