+++ ui/src/shared/components/RefreshingGraph.js
@@ -168,13 +168,22 @@ RefreshingGraph.propTypes = {
   colors: colorsStringSchema,
   cellID: string,
   inView: bool,
-  tableOptions: shape({}),
+  tableOptions: shape({
+    verticalTimeAxis: bool.isRequired,
+    sortBy: shape({
+      internalName: string.isRequired,
+      displayName: string.isRequired,
+      visible: bool.isRequired,
+    }).isRequired,
+    wrapping: string.isRequired,
+    fixFirstColumn: bool.isRequired,
+  }),
   fieldOptions: arrayOf(
     shape({
       internalName: string.isRequired,
       displayName: string.isRequired,
       visible: bool.isRequired,
-    })
+    }).isRequired
   ),
   timeFormat: string.isRequired,
   hoverTime: string.isRequired,
